Move get command message parsers to get package
