+++ src/main/java/seedu/address/logic/commands/GetClubCommand.java
@@ -9,7 +9,7 @@ import seedu.address.commons.util.CopyUtil;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.GetClubMessageParser;
+import seedu.address.logic.parser.get.GetClubMessageParser;
 import seedu.address.model.Model;
 import seedu.address.model.club.Club;
